Fixing multiple-layer bug.
Turns out that if you init a data attribute in a class (i.e. where self.* variables are declared) to a dict(), ALL instances of that class will have _exactly_ the same object bound to the variable name, and hence share data.

So if you have 2 instances, and you modify the dictionary in one it is also updated in the other.

The solution is to push the initilisation into a the __init__() function, and leave a declaration (i.e. "var = None").

Case in point, it was:

	class Preferences:
		local_overlay = dict()
		...

And the fixed version is:

	class Preferences:
		local_overlay = None

		def __init__(self):
			self.local_overlay = dict()

		...

This way the dictionary object is created at the same time as the instance within the scope of the instance, and therefore is unique. Any subsequent instances will have their own scope and so will get their own dictionaries.



git-svn-id: file:///store/svn/tvrenamer@93 0ba0757d-5d28-0410-82a9-afbb3cb19212