Removing dead files and SVN cruft.
Originally this was an SVN repository, and also a private personal playground. This patch removes the SVN cruft (branches / tags directories for now, trunk to follow soon) and a lot of throw-away experimental code which never went anywhere.

(It's a pity I didn't do this clean-up *before* uploading to github!)