Update handling of special episodes listed in EpGuides. Closes #16.
Note that tvrenamer.pl does not consider specials to be in their own season, and so will number the episodes as S1, S2, and so on. This may be confusing if you are accustomed to S01E01, S01E02, and so on as your naming scheme.

Specials are a pain :-)